requirements.txt
# CKD Management RAG System - Dependencies

# =============================================================================
# Core Frameworks
# =============================================================================
langchain>=0.2.0
langgraph>=0.1.0
langchain-huggingface>=0.0.3
langchain-community>=0.2.0
langchain-chroma>=0.1.0

# =============================================================================
# Models & ML
# =============================================================================
transformers>=4.40.0
torch>=2.0.0
accelerate>=0.27.0
bitsandbytes>=0.43.0  # For 4-bit quantization

# =============================================================================
# Embeddings & Vector Store
# =============================================================================
chromadb>=0.4.0
sentence-transformers>=2.6.0

# =============================================================================
# PDF Processing
# =============================================================================
PyMuPDF>=1.24.0
pdfplumber>=0.10.0

# =============================================================================
# PII Detection
# =============================================================================
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.7.0

# =============================================================================
# Evaluation
# =============================================================================
ragas>=0.1.0
langsmith>=0.1.0

# =============================================================================
# UI
# =============================================================================
gradio>=4.0.0

# =============================================================================
# Utilities
# =============================================================================
python-dotenv>=1.0.0
pydantic>=2.0.0
tiktoken>=0.6.0  # For token counting
tqdm>=4.66.0

# =============================================================================
# Development & Testing
# =============================================================================
pytest>=8.0.0
pytest-asyncio>=0.23.0
black>=24.0.0
ruff>=0.3.0
